feat: implement Serde (de)serializer for `Ipld` enum
It's now possible to serialize and deserialize the `Ipld` enum with Serde.
As the Serde data model doesn't support custom types, we use the newtype
struct hack, which is used by many other libraries as well. We introduce
a newtype struct with a mostly unique/hard to clash with name. This struct
wraps the CID.

Serde de(serializers) that support CIDs, can then make use of it, others
will see a strange looking struct in their ouput.