chore: add cargo-release metadata
In order to make releasing easier, add metadata for cargo-release. With this
information you can e.g. run `cargo release --workspace minor` and
it will do the expected thing.

Currently we rely on all crates being bumped to the same version. This is due
to the `shared-version` setting, that makes sure that the commit message has
the `{{version}}` correctly expanded.