feat: implement Serde (de)serializer for `Ipld` enum (#140)
It's now possible to serialize and deserialize the `Ipld` enum with Serde.
As the Serde data model doesn't support custom types, we use the newtype
struct hack, which is used by many other libraries as well. We introduce
a newtype struct with a mostly unique/hard to clash with name. This struct
wraps the CID.

Serde de(serializers) that support CIDs, can then make use of it, others
will see a strange looking struct in their ouput.

This is just a first start, the handling of CIDs still needs to be improved to
make sure that we don't return CIDs when we don't intend to.