fix: improve codec spec compliance
The DAG-CBOR, DAG-JSON and DAG-PB codecs are not more spec compliant. New
negative test fixtures were added to https://github.com/ipld/codec-fixtures,
with this change all pass.

BREAKING CHANGE: DAG-CBOR, DAG-JSON and DAG-PB encoding and decoding is
stricter.

In DAG-CBOR and DAG-JSON now error when duplicated map keys with the same
map are encountered.

DAG-PB is now stricter on encoding and decoding. For example are links and
data not allowed to be interleaved, they have to be one after another (in
any order).

When encoding `Ipld` to DAG-PB, the types now need to match, links need to
be ordered.

The DAG-PB will error if unknown data is encountered during decoding.